Refactor `AddAdminModal` to use standard `Modal` component
